Week 3/Day 2/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
aiofiles==23.2.1
    # via chainlit
aiohttp==3.9.5
    # via
    #   cohere
    #   langchain
    #   langchain-community
    #   python-graphql-client
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==3.7.1
    # via
    #   asyncer
    #   httpcore
    #   openai
    #   starlette
    #   watchfiles
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
asyncer==0.0.2
    # via chainlit
attrs==23.2.0
    # via aiohttp
backoff==2.2.1
    # via cohere
bidict==0.23.1
    # via python-socketio
certifi==2024.6.2
    # via
    #   httpcore
    #   httpx
    #   requests
chainlit==0.7.700
    # via -r requirements.in
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   chainlit
    #   uvicorn
cohere==4.37
    # via -r requirements.in
dataclasses-json==0.5.14
    # via
    #   chainlit
    #   langchain-community
deprecated==1.2.14
    # via
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
distro==1.9.0
    # via openai
exceptiongroup==1.2.1
    # via anyio
fastapi==0.100.1
    # via
    #   chainlit
    #   fastapi-socketio
fastapi-socketio==0.0.10
    # via chainlit
fastavro==1.9.4
    # via cohere
filetype==1.2.0
    # via chainlit
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
googleapis-common-protos==1.63.1
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
grpcio==1.64.1
    # via
    #   grpcio-tools
    #   opentelemetry-exporter-otlp-proto-grpc
    #   qdrant-client
grpcio-tools==1.62.2
    # via qdrant-client
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
    #   wsproto
h2==4.1.0
    # via httpx
hpack==4.0.0
    # via h2
httpcore==0.17.3
    # via httpx
httpx[http2]==0.24.1
    # via
    #   chainlit
    #   openai
    #   qdrant-client
hyperframe==6.0.1
    # via h2
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==6.11.0
    # via
    #   cohere
    #   opentelemetry-api
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.2.4
    # via
    #   -r requirements.in
    #   langchain-community
langchain-community==0.2.4
    # via -r requirements.in
langchain-core==0.2.6
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
    #   langchain-openai
    #   langchain-text-splitters
langchain-openai==0.1.8
    # via -r requirements.in
langchain-text-splitters==0.2.1
    # via langchain
langsmith==0.1.77
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
lazify==0.4.0
    # via chainlit
marshmallow==3.21.3
    # via dataclasses-json
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
nest-asyncio==1.6.0
    # via chainlit
numpy==1.26.4
    # via
    #   langchain
    #   langchain-community
    #   qdrant-client
openai==1.34.0
    # via
    #   -r requirements.in
    #   langchain-openai
opentelemetry-api==1.25.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   uptrace
opentelemetry-exporter-otlp==1.25.0
    # via uptrace
opentelemetry-exporter-otlp-proto-common==1.25.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc==1.25.0
    # via opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http==1.25.0
    # via opentelemetry-exporter-otlp
opentelemetry-instrumentation==0.46b0
    # via uptrace
opentelemetry-proto==1.25.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.25.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   uptrace
opentelemetry-semantic-conventions==0.46b0
    # via opentelemetry-sdk
orjson==3.10.5
    # via langsmith
packaging==23.2
    # via
    #   chainlit
    #   langchain-core
    #   marshmallow
portalocker==2.8.2
    # via qdrant-client
protobuf==4.25.3
    # via
    #   googleapis-common-protos
    #   grpcio-tools
    #   opentelemetry-proto
pydantic==2.7.4
    # via
    #   chainlit
    #   fastapi
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
    #   qdrant-client
pydantic-core==2.18.4
    # via pydantic
pyjwt==2.8.0
    # via chainlit
pypdf2==3.0.1
    # via -r requirements.in
python-dotenv==1.0.0
    # via
    #   -r requirements.in
    #   chainlit
python-engineio==4.9.1
    # via python-socketio
python-graphql-client==0.4.3
    # via chainlit
python-magic==0.4.27
    # via -r requirements.in
python-multipart==0.0.6
    # via chainlit
python-socketio==5.11.2
    # via fastapi-socketio
pyyaml==6.0.1
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
qdrant-client==1.9.1
    # via -r requirements.in
regex==2024.5.15
    # via tiktoken
requests==2.32.3
    # via
    #   cohere
    #   langchain
    #   langchain-community
    #   langsmith
    #   opentelemetry-exporter-otlp-proto-http
    #   python-graphql-client
    #   tiktoken
simple-websocket==1.0.0
    # via python-engineio
sniffio==1.3.1
    # via
    #   anyio
    #   httpcore
    #   httpx
    #   openai
sqlalchemy==2.0.30
    # via
    #   langchain
    #   langchain-community
starlette==0.27.0
    # via fastapi
syncer==2.0.3
    # via chainlit
tenacity==8.3.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tiktoken==0.7.0
    # via
    #   -r requirements.in
    #   langchain-openai
tomli==2.0.1
    # via chainlit
tqdm==4.66.4
    # via openai
typing-extensions==4.12.2
    # via
    #   fastapi
    #   openai
    #   opentelemetry-sdk
    #   pydantic
    #   pydantic-core
    #   pypdf2
    #   sqlalchemy
    #   starlette
    #   typing-inspect
    #   uvicorn
typing-inspect==0.9.0
    # via dataclasses-json
uptrace==1.24.0
    # via chainlit
urllib3==2.2.1
    # via
    #   cohere
    #   qdrant-client
    #   requests
uvicorn==0.23.2
    # via chainlit
watchfiles==0.20.0
    # via chainlit
websockets==12.0
    # via python-graphql-client
wrapt==1.16.0
    # via
    #   deprecated
    #   opentelemetry-instrumentation
wsproto==1.2.0
    # via simple-websocket
yarl==1.9.4
    # via aiohttp
zipp==3.19.2
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
